# contract/openapi.yaml
openapi: 3.1.0
info:
  title: NARAGA API
  version: 2.0.0
  x-contract-version: 2.0.0
  description: |
    Canonical, machine-readable contract for NARAGA — AI-powered geospatial processing.

    THIS FILE IS THE SINGLE SOURCE OF TRUTH. TypeScript types and Pydantic models are
    generated from it; do not hand-edit generated output. CI fails on drift.

    Conventions
    - Every field is snake_case and identical across the Gateway and all services.
    - Timestamps are RFC3339 UTC, millisecond precision, literal `Z`.
    - bbox is ALWAYS EPSG:4326 [min_lon, min_lat, max_lon, max_lat], regardless of `crs`.
    - All models forbid unknown properties, so a typo 422s rather than being silently
      dropped. Standalone objects use additionalProperties: false. Composed models
      (allOf over an OPEN base) use unevaluatedProperties: false at the composed level
      instead — putting additionalProperties: false inside an allOf arm makes the
      schema unsatisfiable in draft 2020-12, because each arm validates in isolation
      and rejects the other arm's properties.
servers:
  - url: http://localhost:8080/v1
    description: Local gateway
...
components:
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Frontend -> Gateway. 15 min TTL.
    internalServiceToken:
      type: http
      scheme: bearer
      description: |
        Gateway <-> Service, both directions. Distinct secret per service AND per
        direction. `kid` header supports dual-key rotation. The gateway MUST verify
        that a callback's calling service matches `job.service`.
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number.
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      description: Items per page. Values above maximum return 422 (never silently clamped).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Required on credit-mutating requests. The original response is replayed for
        24h, so a client retry after a timeout cannot double-debit.
      required: false
      schema: { type: string, format: uuid }
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema: { type: string }
...
  schemas:
...
    Timestamp:
      type: string
      format: date-time
      description: |
        RFC3339 UTC. Producers MUST serialize with millisecond precision and a literal
        `Z` (e.g. 2026-09-01T03:00:00.000Z) — Python's default isoformat() emits
        microseconds and no Z, JS toISOString() emits milliseconds and Z. Timezone-naive
        values are rejected. Consumers MUST accept any valid RFC3339 instant.
        (No regex `pattern` here on purpose: with format=date-time, generators emit a
        real datetime type and a string pattern cannot be applied to it.)
      examples: ["2026-09-01T03:00:00.000Z"]
...
    ServiceName:
      type: string
      enum:
        [
          building_extraction,
          pointcloud_classification,
          landcover_detection,
          reconstruction_3d,
          converter,
        ]
    JobStatus:
      type: string
      enum: [draft, queued, processing, complete, failed, cancelled]
...
    Meta:
      type: object
      required: [request_id, timestamp]
      additionalProperties: false
      properties:
        request_id: { type: string }
        timestamp: { $ref: "#/components/schemas/Timestamp" }
    PageMeta:
      type: object
      description: |
        Deliberately restates Meta's fields instead of allOf-extending it: Meta is
        closed (additionalProperties: false), and a closed schema cannot be extended
        via allOf in draft 2020-12 — the Meta arm would reject the pagination fields.
      required: [request_id, timestamp, total, page, page_size, total_pages]
      additionalProperties: false
      properties:
        request_id: { type: string }
        timestamp: { $ref: "#/components/schemas/Timestamp" }
        total: { type: integer, minimum: 0 }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1 }
        total_pages: { type: integer, minimum: 0 }
    ErrorResponse:
      type: object
      required: [error, meta]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, details]
          additionalProperties: false
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message: { type: string }
            details:
              oneOf: [{ type: object }, { type: "null" }]
        meta: { $ref: "#/components/schemas/Meta" }

    # ---------- entities ----------
...
    BuildingExtractionJob:
      # Closed via unevaluatedProperties at the composed level, NOT
      # additionalProperties: false inside the arm — the arm validates in isolation
      # and would reject every JobBase field, making the schema unsatisfiable.
      # Same idiom on all five Job variants.
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/JobBase"
        - type: object
          required: [service, model, input_dataset_ids, params, result_summary]
          properties:
            service: { type: string, const: building_extraction }
            model: { type: string, enum: [cascade_building, geocarta] }
            input_dataset_ids:
              { $ref: "#/components/schemas/BuildingExtractionInputs" }
            params: { $ref: "#/components/schemas/BuildingExtractionParams" }
            result_summary:
              oneOf:
                [
                  { $ref: "#/components/schemas/BuildingExtractionResult" },
                  { type: "null" },
                ]
    PointCloudClassificationJob:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/JobBase"
        - type: object
          required: [service, model, input_dataset_ids, params, result_summary]
          properties:
            service: { type: string, const: pointcloud_classification }
            model: { type: string, enum: [csf, dgcnn] }
            input_dataset_ids:
              { $ref: "#/components/schemas/PointCloudClassificationInputs" }
            params:
              { $ref: "#/components/schemas/PointCloudClassificationParams" }
            result_summary:
              oneOf:
                [
                  {
                    $ref: "#/components/schemas/PointCloudClassificationResult",
                  },
                  { type: "null" },
                ]
    LandcoverDetectionJob:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/JobBase"
        - type: object
          required: [service, model, input_dataset_ids, params, result_summary]
          properties:
            service: { type: string, const: landcover_detection }
            model: { type: string, enum: [pandala] }
            input_dataset_ids:
              { $ref: "#/components/schemas/LandcoverDetectionInputs" }
            params: { $ref: "#/components/schemas/LandcoverDetectionParams" }
            result_summary:
              oneOf:
                [
                  { $ref: "#/components/schemas/LandcoverDetectionResult" },
                  { type: "null" },
                ]
    Reconstruction3dJob:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/JobBase"
        - type: object
          required: [service, model, input_dataset_ids, params, result_summary]
          properties:
            service: { type: string, const: reconstruction_3d }
            model: { type: string, enum: [dream3d, cascade_mesh] }
            input_dataset_ids:
              { $ref: "#/components/schemas/Reconstruction3dInputs" }
            params: { $ref: "#/components/schemas/Reconstruction3dParams" }
            result_summary:
              oneOf:
                [
                  { $ref: "#/components/schemas/Reconstruction3dResult" },
                  { type: "null" },
                ]
    ConverterJob:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/JobBase"
        - type: object
          required: [service, model, input_dataset_ids, params, result_summary]
          properties:
            service: { type: string, const: converter }
            model:
              type: "null"
              description: |
                Converter has no model. The PDF made Job.model a required non-null
                string while listing no models for converter, so every converter job
                had to send a value that could not validate (CONTRACT-CHANGES A7).
            input_dataset_ids: { $ref: "#/components/schemas/ConverterInputs" }
            params: { $ref: "#/components/schemas/ConverterParams" }
            result_summary:
              oneOf:
                [
                  { $ref: "#/components/schemas/ConverterResult" },
                  { type: "null" },
                ]
    Job:
      oneOf:
        - $ref: "#/components/schemas/BuildingExtractionJob"
        - $ref: "#/components/schemas/PointCloudClassificationJob"
        - $ref: "#/components/schemas/LandcoverDetectionJob"
        - $ref: "#/components/schemas/Reconstruction3dJob"
        - $ref: "#/components/schemas/ConverterJob"
      discriminator:
        propertyName: service
        mapping:
          building_extraction: "#/components/schemas/BuildingExtractionJob"
          pointcloud_classification: "#/components/schemas/PointCloudClassificationJob"
          landcover_detection: "#/components/schemas/LandcoverDetectionJob"
          reconstruction_3d: "#/components/schemas/Reconstruction3dJob"
          converter: "#/components/schemas/ConverterJob"

    # ---------- callback ----------
...
    CreateJobRequest:
      description: |
        Discriminated on `service` so `params`, `model`, and `input_dataset_ids` are
        validated together. All referenced datasets MUST be status=ready at both create
        and submit (422 DATASET_NOT_READY otherwise) — a draft may be submitted days later.
      oneOf:
        - $ref: "#/components/schemas/CreateBuildingExtractionJobRequest"
        - $ref: "#/components/schemas/CreatePointCloudClassificationJobRequest"
        - $ref: "#/components/schemas/CreateLandcoverDetectionJobRequest"
        - $ref: "#/components/schemas/CreateReconstruction3dJobRequest"
        - $ref: "#/components/schemas/CreateConverterJobRequest"
      discriminator:
        propertyName: service
        mapping:
          building_extraction: "#/components/schemas/CreateBuildingExtractionJobRequest"
          pointcloud_classification: "#/components/schemas/CreatePointCloudClassificationJobRequest"
          landcover_detection: "#/components/schemas/CreateLandcoverDetectionJobRequest"
          reconstruction_3d: "#/components/schemas/CreateReconstruction3dJobRequest"
          converter: "#/components/schemas/CreateConverterJobRequest"
...
    CreateBuildingExtractionJobRequest:
      allOf:
        - $ref: "#/components/schemas/CreateJobRequestBase"
        - type: object
          required: [service, model, input_dataset_ids, params]
          additionalProperties: false
          properties:
            service: { type: string, const: building_extraction }
            model: { type: string, enum: [cascade_building, geocarta] }
            job_name: { type: string }
            save_as_draft: { type: boolean }
            input_dataset_ids:
              { $ref: "#/components/schemas/BuildingExtractionInputs" }
            params: { $ref: "#/components/schemas/BuildingExtractionParams" }
    CreatePointCloudClassificationJobRequest:
      allOf:
        - $ref: "#/components/schemas/CreateJobRequestBase"
        - type: object
          required: [service, model, input_dataset_ids, params]
          additionalProperties: false
          properties:
            service: { type: string, const: pointcloud_classification }
            model: { type: string, enum: [csf, dgcnn] }
            job_name: { type: string }
            save_as_draft: { type: boolean }
            input_dataset_ids:
              { $ref: "#/components/schemas/PointCloudClassificationInputs" }
            params:
              { $ref: "#/components/schemas/PointCloudClassificationParams" }
    CreateLandcoverDetectionJobRequest:
      allOf:
        - $ref: "#/components/schemas/CreateJobRequestBase"
        - type: object
          required: [service, model, input_dataset_ids, params]
          additionalProperties: false
          properties:
            service: { type: string, const: landcover_detection }
            model: { type: string, enum: [pandala] }
            job_name: { type: string }
            save_as_draft: { type: boolean }
            input_dataset_ids:
              { $ref: "#/components/schemas/LandcoverDetectionInputs" }
            params: { $ref: "#/components/schemas/LandcoverDetectionParams" }
    CreateReconstruction3dJobRequest:
      allOf:
        - $ref: "#/components/schemas/CreateJobRequestBase"
        - type: object
          required: [service, model, input_dataset_ids, params]
          additionalProperties: false
          properties:
            service: { type: string, const: reconstruction_3d }
            model: { type: string, enum: [dream3d, cascade_mesh] }
            job_name: { type: string }
            save_as_draft: { type: boolean }
            input_dataset_ids:
              { $ref: "#/components/schemas/Reconstruction3dInputs" }
            params: { $ref: "#/components/schemas/Reconstruction3dParams" }
    CreateConverterJobRequest:
      allOf:
        - $ref: "#/components/schemas/CreateJobRequestBase"
        - type: object
          required: [service, input_dataset_ids, params]
          additionalProperties: false
          properties:
            service: { type: string, const: converter }
            model: { type: "null" }
            job_name: { type: string }
            save_as_draft: { type: boolean }
            input_dataset_ids: { $ref: "#/components/schemas/ConverterInputs" }
            params: { $ref: "#/components/schemas/ConverterParams" }
    EstimateResponse:
      type: object
      required: [credits_estimated, estimated_output_bytes, breakdown]
      additionalProperties: false
      properties:
        credits_estimated: { type: integer, minimum: 0 }
        estimated_output_bytes: { type: integer, minimum: 0, format: int64 }
        breakdown:
          type: object
          required: [line_items, total]
          additionalProperties: false
          properties:
            line_items:
              type: array
              items:
                type: object
                required: [label, unit, quantity, unit_credits, subtotal]
                additionalProperties: false
                properties:
                  label: { type: string }
                  unit: { type: string }
                  quantity: { type: number }
                  unit_credits: { type: number }
                  subtotal: { type: integer }
            total: { type: integer }
...
  responses:
    BadRequest:
      description: Validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFound:
      description: |
        Not found, OR found but owned by another user — cross-owner access returns 404
        rather than 403 to prevent ID enumeration (CONTRACT-CHANGES C2).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: State conflict (duplicate job, illegal transition, dataset in use).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    PaymentRequired:
      description: Insufficient credits.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnprocessableEntity:
      description: Semantically invalid (bad model/dataset combination, missing input, CRS mismatch).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnexpectedError:
      description: |
        Any status not otherwise documented — including 500 INTERNAL_ERROR. Every
        operation carries this as `default` so clients always have a typed error body
        to fall back on instead of parsing ad-hoc text.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
security:
  - accessToken: []
paths:
...
  /jobs:
    get:
      tags: [jobs]
      summary: List jobs
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - name: service
          in: query
          schema:
            type: array
            items: { $ref: "#/components/schemas/ServiceName" }
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: "#/components/schemas/JobStatus" }
        - name: job_ids
          in: query
          description: Bulk fetch so a dashboard polls once, not once per job.
          schema:
            type: array
            items: { type: string, format: uuid }
      responses:
        default: { $ref: "#/components/responses/UnexpectedError" }
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Job" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [jobs]
      summary: Create a job (draft by default)
      description: |
        With save_as_draft=false the job is queued AND credits are debited immediately,
        so this can return 402. Validates the full per-service shape, including 3D
        Reconstruction's elevation_source -> required-inputs rule.
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKey" }
        - { $ref: "#/components/parameters/RequestId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateJobRequest" }
      responses:
        default: { $ref: "#/components/responses/UnexpectedError" }
        "201":
          description: Created.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                additionalProperties: false
                properties:
                  data:
                    type: object
                    required: [job]
                    additionalProperties: false
                    properties:
                      job: { $ref: "#/components/schemas/Job" }
                  meta: { $ref: "#/components/schemas/Meta" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/PaymentRequired" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /jobs/{job_id}:
    parameters:
      - name: job_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [jobs]
      summary: Get a job
      description: Supports ETag / If-None-Match so progress polling is cheap.
      responses:
        default: { $ref: "#/components/responses/UnexpectedError" }
        "200":
          description: OK
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                additionalProperties: false
                properties:
                  data:
                    type: object
                    required: [job]
                    additionalProperties: false
                    properties:
                      job: { $ref: "#/components/schemas/Job" }
                  meta: { $ref: "#/components/schemas/Meta" }
        "304": { description: Not modified. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [jobs]
      summary: Edit a draft job
      description: |
        Drafts only (409 otherwise). `params` is REPLACED wholesale, never deep-merged.
        Re-runs full cross-field validation and recomputes credits_estimated.
        `service` is immutable. Supports If-Match for optimistic concurrency.
      parameters:
        - name: If-Match
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateJobRequest" }
      responses:
        default: { $ref: "#/components/responses/UnexpectedError" }
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                additionalProperties: false
                properties:
                  data:
                    type: object
                    required: [job]
                    additionalProperties: false
                    properties:
                      job: { $ref: "#/components/schemas/Job" }
                  meta: { $ref: "#/components/schemas/Meta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { description: Precondition failed (concurrent edit). }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      tags: [jobs]
      summary: Delete a draft job
      description: Drafts only; 409 JOB_NOT_DELETABLE otherwise.
      responses:
        default: { $ref: "#/components/responses/UnexpectedError" }
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /jobs/{job_id}/estimate:
    post:
      tags: [jobs]
      summary: Estimate credits
      description: |
        Proxies the owning service's /estimate so pricing lives in exactly one place —
        the PDF returned credits_estimated from three places with no authority, inviting
        duplicate pricing logic in TypeScript and Python (CONTRACT-CHANGES D9).
        Persists credits_estimated on the job, which is why this is a POST.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        default: { $ref: "#/components/responses/UnexpectedError" }
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                additionalProperties: false
                properties:
                  data: { $ref: "#/components/schemas/EstimateResponse" }
                  meta: { $ref: "#/components/schemas/Meta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...